+++ selendroid-server/src/main/java/io/selendroid/server/handler/SwitchWindow.java
@@ -14,14 +14,14 @@
 package io.selendroid.server.handler;
 
 import io.selendroid.android.WindowType;
+import io.selendroid.exceptions.SelendroidException;
 import io.selendroid.server.RequestHandler;
 import io.selendroid.server.Response;
-
-import org.json.JSONException;
-import io.selendroid.exceptions.SelendroidException;
 import io.selendroid.server.SelendroidResponse;
 import io.selendroid.server.model.SelendroidDriver;
 import io.selendroid.util.SelendroidLogger;
+
+import org.json.JSONException;
 import org.webbitserver.HttpRequest;
 
 public class SwitchWindow extends RequestHandler {
@@ -35,13 +35,13 @@ public class SwitchWindow extends RequestHandler {
     SelendroidLogger.log("Switch Window command");
     String windowName = getPayload().getString("name");
     if (windowName == null || windowName.isEmpty()) {
-      return new SelendroidResponse(getSessionId(), 13, new SelendroidException("Window name is missing."));
+      return new SelendroidResponse(getSessionId(), 13, new SelendroidException(
+          "Window name is missing."));
     }
     SelendroidDriver driver = getSelendroidDriver();
-    if (WindowType.NATIVE_APP.name().equals(windowName)) {
-      driver.switchDriverMode(WindowType.NATIVE_APP);
-    } else if (WindowType.WEBVIEW.name().equals(windowName)) {
-      driver.switchDriverMode(WindowType.WEBVIEW);
+    if (windowName.startsWith(WindowType.NATIVE_APP.name())
+        || windowName.startsWith(WindowType.WEBVIEW.name())) {
+      driver.switchDriverMode(windowName);
     } else {
       return new SelendroidResponse(getSessionId(), 23, new SelendroidException(
           "Invalid window handle was used: only 'NATIVE_APP' and 'WEBVIEW' are supported."));